maps fragment acabado

+++ app/src/main/java/com/example/politravel_mariomolina/Paquet.kt
@@ -2,5 +2,5 @@ package com.example.politravel_mariomolina
 
 import java.io.Serializable
 
-class Paquet(val idPaquet: Int,val nomPaquet: String,val descripcio: String, val imgLlista: String,val imgDetail: String,val transport:String, val iniciRecorregut: String , val longitud: Double , val latitud: Double, val grausGoogleMaps: Int,val finalRecorregut: String,val dies: Int,val itinerari: MutableList<Itinerari>):
+class Paquet(val idPaquet: Int,val nomPaquet: String,val descripcio: String, val imgLlista: String,val imgDetail: String,val transport:String, val iniciRecorregut: String , val longitud: Double , val latitud: Double, val grausGoogleMaps: Float,val finalRecorregut: String,val dies: Int,val itinerari: MutableList<Itinerari>):
     Serializable
